feat: 매장 등록 현황 리스트 UI 구현
